Missed a file in last commit
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Rules/APP@5152 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ model/src/main/java/eu/europa/ec/fisheries/uvms/rules/model/dto/AlarmListResponseDto.java
@@ -1,7 +1,6 @@
 package eu.europa.ec.fisheries.uvms.rules.model.dto;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.List;
 
 import eu.europa.ec.fisheries.schema.rules.alarm.v1.AlarmType;
@@ -11,8 +10,8 @@ public class AlarmListResponseDto implements Serializable {
     private static final long serialVersionUID = 1;
 
     private List<AlarmType> alarmList;
-    private BigInteger totalNumberOfPages;
-    private BigInteger currentPage;
+    private int totalNumberOfPages;
+    private int currentPage;
 
     public List<AlarmType> getAlarmList() {
         return alarmList;
@@ -22,19 +21,19 @@ public class AlarmListResponseDto implements Serializable {
         this.alarmList = alarmList;
     }
 
-    public BigInteger getTotalNumberOfPages() {
+    public int getTotalNumberOfPages() {
         return totalNumberOfPages;
     }
 
-    public void setTotalNumberOfPages(BigInteger totalNumberOfPages) {
+    public void setTotalNumberOfPages(int totalNumberOfPages) {
         this.totalNumberOfPages = totalNumberOfPages;
     }
 
-    public BigInteger getCurrentPage() {
+    public int getCurrentPage() {
         return currentPage;
     }
 
-    public void setCurrentPage(BigInteger currentPage) {
+    public void setCurrentPage(int currentPage) {
         this.currentPage = currentPage;
     }
